requirements.txt
langchain==0.3.21
psycopg2-binary==2.9.10
dotenv-python==0.0.1
langchain-community==0.3.20
llama-index-llms-huggingface-api==0.4.1
langchain-openai==0.3.11
langgraph==0.3.21